Added Fingerprint Protection setting disclaimer on installation
This adds a disclaimer that only shows on initial installation. This
disclaimer informs the user about the pros and cons of the Fingerprint
Protection setting and allows the user to choose whether to turn it on
or disable it from that point forward. The value of the setting can
still be changed from the "Extension Settings" panel.