Cleaned up css and coding style
This commit ties up items missed from the "no-regions" branch that were
missed but found on the port over to Chrome. They mainly have to do with
coding style but a few items deal with minor CSS enhancements and one
component not being fully moved over but still fully working.